Moved all the movment variables into a singular byte in Entity
combined my uint8_t that held the input variables WASD and the bools and short from Entity into a singular uint8_t in Entity to save two bytes on the two bools, removed the sealing check bool cause it's only checked a maximum of once anyways, and saved two more bytes on the short, saving me a total of 5 bytes. And hopefully not loosing any preformence speed, unless we even gained some.